Only consider complete parcels when determining if a dataset is due.
For example, consider a dataset whose interval is "1 hour", and for which
the only parcel less than an hour old is incomplete.

Previously, this dataset would not have been considered due, and running
`export_manager process` on it would not have triggered an export.

Now, it will be considered due, because there is no _complete_ parcel
within the interval. So `export_manager process` will trigger an export.

This is helpful for exports with very long intervals, so that you don't
go days or weeks without another export when the previous one failed.

On the other hand, it would problematic if you are running
`export_manager process` so frequently that new runs begin before old
runs finish, since you could end up doing multiple exports concurrently.
My current advice is "don't do that", although changes to better handle
that scenario could be considered in the future.